Factor out a "FV_telephone()" method, to validate phone numbers for any country.
While "FV_american_phone()" is nice, its not of use when wanting to
validate numbers for other countries.  So, factor out a "FV_telephone()"
method which requires a list of country codes (e.g. "US", "CA", "UK") to
be passed in and which it can then use to validate the telephone number
to ensure that its valid in one of those countries.